.ipynb_checkpoints/requirements-checkpoint.txt
# Data processing
pandas==2.0.3
numpy==1.24.3

# Dashboard
dash==2.14.0
plotly==5.17.0
dash-bootstrap-components==1.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
gunicorn==20.1.0